# [Feature] bulk load optimization(Part 3 Block merger)