## Why I'm doing:

## What I'm doing:

Implement block group merge in SpillMemTableSink

Fixes #53954

```
Note: Google Test filter = SpillMemTableSinkTest.test_merge
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from SpillMemTableSinkTest
[ RUN      ] SpillMemTableSinkTest.test_merge
I20241225 11:26:56.020481 137329194452928 spill_mem_table_sink.cpp:181] LogBlock[container=/tmp/.org.chromium.Chromium.kLhTrC/spill/00000000-0000-0000-0000-000000000000/00000000-0000-0001-0000-000000000001-load_spill-0-0]
I20241225 11:26:56.020657 137329194452928 spill_mem_table_sink.cpp:181] LogBlock[container=/tmp/.org.chromium.Chromium.kLhTrC/spill/00000000-0000-0000-0000-000000000000/00000000-0000-0001-0000-000000000001-load_spill-0-0]
I20241225 11:26:56.023854 137329194452928 spill_mem_table_sink.cpp:242] SpillMemTableSink merge finished, txn:2 tablet:1 blockgroups:1 blocks:2 input_bytes:8192 merges:1 rows:24 chunks:1 duration:3ms
[       OK ] SpillMemTableSinkTest.test_merge (1001 ms)
[----------] 1 test from SpillMemTableSinkTest (1001 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test suite ran. (1002 ms total)
[  PASSED  ] 1 test.
I20241225 11:26:57.021432 137326344799936 runtime_filter_worker.cpp:1004] RuntimeFilterWorker going to exit.
I20241225 11:26:57.097586 137326298662592 profile_report_worker.cpp:124] ProfileReportWorker going to exit.
I20241225 11:26:57.103712 137329194452928 pipeline_executor_set.cpp:152] [WORKGROUP] close executors ([name=com] [num_driver_threads=2] [num_scan_threads=2] [num_connector_scan_threads=24] [cpuids=(0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23)] [conf=([num_total_cores=24] [num_total_driver_threads=2] [num_total_scan_threads=2] [num_total_connector_scan_threads=24] [enable_bind_cpus=true] [enable_cpu_borrowing=true])])
I20241225 11:26:57.197936 137326275593920 result_buffer_mgr.cpp:194] result buffer manager cancel thread finish.
I20241225 11:26:57.488567 137328121087680 fragment_mgr.cpp:552] FragmentMgr cancel worker is going to exit.

```
## What type of PR is this:

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.4
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0